[과제] 3주차 과제 업로드(수정) - #10

+++ dongmin/src/main/java/dongmin/code/dongmin/domain/user/service/UserService.java
@@ -3,6 +3,7 @@ package dongmin.code.dongmin.domain.user.service;
 import dongmin.code.dongmin.domain.user.dto.UserCreateRequestDTO;
 import dongmin.code.dongmin.domain.user.dto.UserResponseDTO;
 import dongmin.code.dongmin.domain.user.entity.User;
+import dongmin.code.dongmin.domain.user.entity.UserDetailsImpl;
 import dongmin.code.dongmin.domain.user.repository.UserRepository;
 import dongmin.code.dongmin.global.exception.error.RestApiException;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
-import static dongmin.code.dongmin.global.exception.error.CustomErrorCode.USER_NOT_FOUND;
+import static dongmin.code.dongmin.global.exception.error.CustomErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -41,26 +42,27 @@ public class UserService {
     }
 
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, UserDetailsImpl userDetails) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RestApiException(USER_NOT_FOUND));
+
+        if(!id.equals(userDetails.getUser().getId())){
+            throw new RestApiException(DELETE_FORBIDDEN);
+        }
 
         userRepository.deleteById(id);
     }
 
     @Transactional
-    public void update(Long id, UserCreateRequestDTO userCreateRequestDTO) {
+    public void update(Long id, UserCreateRequestDTO userCreateRequestDTO, UserDetailsImpl userDetails) {
+
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RestApiException(USER_NOT_FOUND));
 
-        user.update(
-                userCreateRequestDTO.getName(),
-                userCreateRequestDTO.getEmail(),
-                bCryptPasswordEncoder.encode(userCreateRequestDTO.getPassword()),
-                userCreateRequestDTO.getPart(),
-                userCreateRequestDTO.getGen(),
-                userCreateRequestDTO.getPhoneNumber(),
-                userCreateRequestDTO.getJoinDate()
-        );
+        if(!id.equals(userDetails.getUser().getId())){
+            throw new RestApiException(CHANGE_FORBIDDEN);
+        }
+
+        user.update(userCreateRequestDTO, bCryptPasswordEncoder.encode(userCreateRequestDTO.getPassword()));
     }
 }
